requirements.txt
alembic==0.9.3.dev0
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.4.2
Babel==2.4.0
beautifulsoup4==4.9.1
blinker==1.4
Brlapi==0.6.6
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
crcmod==1.7
cryptography==2.1.4
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
decorator==4.1.2
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
editdistance==0.5.3
filelock==3.0.12
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.15.2
GDAL==3.0.4
html5lib==0.999999999
httplib2==0.9.2
idna==2.10
importlib-metadata==1.6.0
importlib-resources==1.4.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
langdetect==1.0.8
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
ldap3==2.4.1
louis==3.5.0
lxml==4.5.2
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
nbformat==4.4.0
neo4j==4.1.0
netifaces==0.10.4
numpy==1.19.1
oauth==1.0.1
olefile==0.45.1
opencv-contrib-python==4.2.0.34
OWSLib==0.18.0
pandas==1.1.0
paramiko==2.0.0
passlib==1.7.1
pbr==5.4.5
pexpect==4.2.1
Pillow==7.2.0
plotly==2.2.3
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.8.5
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pyinotify==0.9.6
pylint==2.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyproj==2.5.0
pyRFC3339==1.0
pytesseract==0.3.5
pytesstrain==0.1.15
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.13.2
six==1.15.0
soupsieve==2.0.1
speaklater==1.4
SQLAlchemy==1.1.11
sqlparse==0.2.4
ssh-import-id==5.7
sshtunnel==0.1.4
stevedore==1.32.0
system-service==0.3
systemd-python==234
tesseract==0.1.3
tesseract-ocr==0.0.1
toml==0.10.1
traitlets==4.3.2
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.3
virtualenv==20.0.17
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
